Separate specs between homepage and results page
  Our system tests are unweildy. I've tried to separate them into
  smaller more specific scenarios.

  One apparent discrimintaion I can make is to separate specs that
  happen on the home page and those that happen on the results page.
