+++ spec/system/find/results_page/view_course_spec.rb
